list: rename Value to value, len uint to int
Two cosmetic follow-ups from the type-design review:

* element.Value → element.value. The upper-case V was a
  carry-over from container/list where the field is exported for
  external iteration. Nothing outside this package needs it, so
  lowercasing aligns it with the (already-private) surrounding
  types.

* list.len uint → list.len int. The unsigned type was originally
  load-bearing against underflow in the pre-hardening remove path;
  now that remove early-returns when e.list != l, len can never
  underflow from ordinary use. int is the conventional Go choice
  and propagates to WaitersLen/waitForWaiters signatures more
  naturally (no uint(i+1) casts in tests).

No behavior change. All tests pass under -race.

+++ fifomu_context_test.go
@@ -267,7 +267,7 @@ func TestLockContext_HammerConcurrent(t *testing.T) {
 // waitForWaiters blocks until m has at least n queued waiters, or
 // fails the test after a generous deadline. Replaces sleep-based
 // gating so the FIFO tests don't flake on loaded CI.
-func waitForWaiters(t *testing.T, m *fifomu.Mutex, n uint) {
+func waitForWaiters(t *testing.T, m *fifomu.Mutex, n int) {
 	t.Helper()
 	deadline := time.Now().Add(2 * time.Second)
 	for fifomu.WaitersLen(m) < n {
@@ -297,7 +297,7 @@ func TestMutex_FIFOOrdering(t *testing.T) {
 			order <- i
 			mu.Unlock()
 		}(i)
-		waitForWaiters(t, &mu, uint(i+1))
+		waitForWaiters(t, &mu, i+1)
 	}
 
 	mu.Unlock()
@@ -337,7 +337,7 @@ func TestLockContext_FIFOAmongContextWaiters(t *testing.T) {
 			order <- i
 			mu.Unlock()
 		}(i)
-		waitForWaiters(t, &mu, uint(i+1))
+		waitForWaiters(t, &mu, i+1)
 	}
 
 	mu.Unlock()
